% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRP.R
\name{BRP_read}
\alias{BRP_read}
\title{Read BRP Data.}
\usage{
BRP_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename of the BRP data.}
}
\value{
This function returns a tibble containing the BRP data. If an
  incorrect filename is entered the function will stop.
}
\description{
\code{BRP_read} reads BRP data into the environment.
}
\details{
This function reads data from the  US National Highway Traffic Safety
  Administration's Fatality Analysis Reporting System (BRP), given a
  filename for the data. It returns a tibble of the data. For this function
  to work properly, a filename pointing to an existing file must be given.
}
\examples{
BRP_read(filename = 'accident_2013.csv.bz2')

\dontrun{
BRP_read(filename = 'filedoesnotexist')
}

}
