% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRP.R
\name{BRP_read_years}
\alias{BRP_read_years}
\title{Read BRP files for one or more years.}
\usage{
BRP_read_years(years)
}
\arguments{
\item{years}{Vector of years' BRP files to open.  Vector members must be
an integer, or a string or numeric that can be coerced to a string,
of the year of interest.}
}
\value{
This function returns a list of tibbles, where each tibble contains
  containing the year and month from the observations in the corresponding
  year's BRP data.  If an invalid year is given, the corresponding
  list will be NULL.
}
\description{
\code{BRP_read_years} produces a list of tibbles of BRP data, given an
  input vector of years.
}
\details{
This function takes a vector of years and produces a list of tibbles,
  where each tibble is that year's BRP file year and MONTH observations.
  This is a simple function that strips all useful data out of the BRP
  tables and produces a completely useless tibble, but is meant for
  practice in the Coursera course.  For this function to work, valid
  years should be entered. Invalid years will have a NULL entry in the
  returned list.
}
\examples{
BRP_read_years(years = c(2013, 2014, 2015))
BRP_read_years(years = 2013)

\dontrun{
BRP_read_years(years = 2000) # error
}

}
