% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRP.R
\name{BRP_summarize_years}
\alias{BRP_summarize_years}
\title{Produce a Summary of BRP Files.}
\usage{
BRP_summarize_years(years)
}
\arguments{
\item{years}{Vector of years' BRP files to open.  Vector members must be
an integer, or a string or numeric that can be coerced to a string.}
}
\value{
This function returns tibble where the first column is the month,
  the second and following columns are the requested years, and the
  rows for the year columns are the number of BRP observations for
  that month/year combination.  The returned columns are only for years
  with valid BRP data. If no valid years are found, the function
  well error out.
}
\description{
\code{BRP_summarize_years} produces a summary tibble of BRP years and
  months given a vector of years.
}
\details{
This function takes a vector of years, pulls the BRP data for
  those years, and then produces a summary tibble. The summary tibble shows
  the number of observations for each month/year combination for the
  extracted BRP data. For this function to work properly, the years must
  be years with valid data.
}
\examples{
BRP_summarize_years(years = c(2013, 2014, 2015))
BRP_summarize_years(years = 2013)

\dontrun{
BRP_summarize_years(years = 2000)
}

}
